+++ CHANGELOG.md
@@ -16,6 +16,7 @@ but cannot always guarantee backwards compatibility. Changes that may **break co
 **Dependencies**
 
 ### For developers of the library:
+- Refactored single-vs-sequence returns across the library to uniformly use `series2seq`, instead of per-site hardcoded `isinstance(series, TimeSeries)`. Behavior preserving groundwork for the upcoming `TimeSeriesSequence` migration ([#3037](https://github.com/unit8co/darts/issues/3037)). [#3131](https://github.com/unit8co/darts/pull/3131) by [Oswald Zink](https://github.com/ozink-u8).
 
 ## [0.45.0](https://github.com/unit8co/darts/tree/0.45.0) (2026-06-19)